feat: create useSearchUser hook
